requirements.txt
Pillow
deezer-python
moviepy
google-trans-new
nospamplus
eyed3
img2pdf
pymediainfo
apscheduler
git+https://github.com/chsaiujwal/maigret
git+https://github.com/gleitz/howdoi
git+https://github.com/tushar-rishav/code2pdf
pyseoanalyzer
fake_useragent
pygifsicle
pokedex.py
faker
tracemoepy
deep_translator
countryinfo
mal-api
youtube-search-python
emoji-country-flag
langdetect
PyProxyToolkit
cairosvg
grapheme
cryptosteganography
QScintilla
requests_html
flask
yahoo_fin
anime_downloader
pandas
fontTools
pydub
quote
lottie
textblob
python-magic
glitch_this
PyGithub
pytesseract
youtube-dl
opencv-contrib-python
Telethon==1.19.0
iplookup
wget
numpy
pyshorteners
aiohttp
bs4
coffeehouse
cowpy
emoji==0.6.0
gTTS-token>=1.1.3
gTTS>=2.0.1
geopy
gitpython
google-api-python-client==1.8.0
google-auth-oauthlib
gsearch
hachoir
heroku3
httplib2
humanize
lxml
lyricsgenius
oauth2client
psycopg2-binary
PyLyrics
pySmartDL
pybase64>=0.4.0
pyfiglet
pylast
python-barcode
python-dotenv
pytz
qrcode
regex
requests
search-engine-parser>=0.4.2
selenium
speedtest-cli>=2.0.2
sqlalchemy>=1.2
telegraph
tswift
urbandict>=0.5
wikipedia>=1.4.0
youtube-search
password_strength
pyjokes
cryptocompare==0.7.3
googletrans
pdf2docx
